Fix bug that user could remove members from another garden, end tests for removUser, deleteGarden, getGarden, createGarden, getGardens TG-198 #done TG-121 #done TG-123 #done TG-221 #done

+++ src/main/java/com/potus/app/garden/controller/GardenController.java
@@ -141,8 +141,14 @@ public class GardenController {
 
         User userRequest = userService.findByUsername(user);
         GardenMember memberRequest = gardenService.findByUser(userRequest);
+
+        if(memberRequest.getGarden() != selectedGarden)
+            throw new ResourceNotFoundException(USER_NOT_MEMBER);
+
         if(member.getRole().compareTo(memberRequest.getRole()) < 1)
             throw new ForbiddenException();
+
+
 
         gardenService.removeUser(memberRequest);
     }
